fix: 修复库存数量类型不匹配问题
1. 将库存服务中的数量类型从 Integer 改为 Long
2. 在入库和出库服务中使用 Long.valueOf() 进行类型转换
3. 优化库存操作的错误处理

+++ ruoyi-system/src/main/java/com/ruoyi/system/tile/service/ITileStockService.java
@@ -66,7 +66,7 @@ public interface ITileStockService
      * @param quantity 增加数量
      * @return 结果
      */
-    public int addStock(Long goodsId, Long warehouseId, Integer quantity);
+    public int addStock(Long goodsId, Long warehouseId, Long quantity);
 
     /**
      * 减少库存
@@ -76,5 +76,5 @@ public interface ITileStockService
      * @param quantity 数量
      * @return 结果
      */
-    public int subtractStock(Long goodsId, Long warehouseId, Integer quantity);
+    public int subtractStock(Long goodsId, Long warehouseId, Long quantity);
 }
